+++ src/drto/__init__.py
@@ -30,6 +30,7 @@ from drto.declarations import (
 )
 from drto.info import Info, info
 from drto.initialize_steady_state import SteadyStateInitReport, initialize_steady_state
+from drto.cold_start import ColdStartReport, cold_start_dynamic
 from drto.objective import build_objective
 from drto.plotting import plot_controls, plot_stage_cost, plot_states
 
@@ -51,6 +52,8 @@ __all__ = [
     "Info",
     "info",
     "build_objective",
+    "cold_start_dynamic",
+    "ColdStartReport",
     "plot_states",
     "plot_controls",
     "plot_stage_cost",